Django-Middleware-0x03/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-filter>=23.0
python-decouple>=3.8
django-cors-headers>=4.0
psycopg2-binary>=2.9
pytest>=8.0
pytest-django>=4.8
flake8>=6.1
black>=24.0
pip-audit>=2.7
drf-spectacular>=0.27


